+++ apps/backend/src/app/api/latest/emails/send-email/route.tsx
@@ -1,7 +1,7 @@
 import { renderEmailWithTheme } from "@/lib/email-themes";
 import { getEmailConfig, sendEmail } from "@/lib/emails";
 import { getNotificationCategoryByName, hasNotificationEnabled } from "@/lib/notification-categories";
-import { prismaClient } from "@/prisma-client";
+import { getPrismaClientForTenancy } from "@/prisma-client";
 import { createSmartRouteHandler } from "@/route-handlers/smart-route-handler";
 import { adaptSchema, serverOrHigherAuthTypeSchema, yupArray, yupBoolean, yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
 import { getEnvVariable } from "@stackframe/stack-shared/dist/utils/env";
@@ -56,8 +56,13 @@ export const POST = createSmartRouteHandler({
     if (!notificationCategory) {
       throw new StatusError(404, "Notification category not found");
     }
+    const themeList = auth.tenancy.completeConfig.emails.themeList;
+    if (!Object.keys(themeList).includes(auth.tenancy.completeConfig.emails.theme)) {
+      throw new StatusError(400, "No active theme found");
+    }
+    const activeTheme = themeList[auth.tenancy.completeConfig.emails.theme];
 
-    const users = await prismaClient.projectUser.findMany({
+    const users = await getPrismaClientForTenancy(auth.tenancy).projectUser.findMany({
       where: {
         tenancyId: auth.tenancy.id,
         projectUserId: {
@@ -78,7 +83,7 @@ export const POST = createSmartRouteHandler({
         userSendErrors.set(userId, "User not found");
         continue;
       }
-      const isNotificationEnabled = await hasNotificationEnabled(auth.tenancy.id, user.projectUserId, notificationCategory.id);
+      const isNotificationEnabled = await hasNotificationEnabled(auth.tenancy, user.projectUserId, notificationCategory.id);
       if (!isNotificationEnabled) {
         userSendErrors.set(userId, "User has disabled notifications for this category");
         continue;
@@ -107,7 +112,12 @@ export const POST = createSmartRouteHandler({
         unsubscribeLink = unsubUrl.toString();
       }
 
-      const renderedEmail = await renderEmailWithTheme(body.html, auth.tenancy.config.email_theme, unsubscribeLink);
+
+      const renderedEmail = await renderEmailWithTheme(
+        body.html,
+        activeTheme.tsxSource,
+        unsubscribeLink || undefined
+      );
       if ("error" in renderedEmail) {
         userSendErrors.set(userId, "There was an error rendering the email");
         continue;