+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @stackframe/stack-backend
 
+## 2.8.21
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.21
+  - @stackframe/stack-emails@2.8.21
+
+## 2.8.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.20
+  - @stackframe/stack-emails@2.8.20
+
 ## 2.8.19
 
 ### Patch Changes