+++ apps/mock-oauth-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/mock-oauth-server
 
+## 2.8.21
+
+## 2.8.20
+
 ## 2.8.19
 
 ## 2.8.18