+++ docs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackframe/stack-docs
 
+## 2.8.21
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack@2.8.21
+
+## 2.8.20
+
+### Patch Changes
+
+- @stackframe/stack@2.8.20
+
 ## 2.8.19
 
 ### Patch Changes