+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/js-example
 
+## 2.8.21
+
+### Patch Changes
+
+- @stackframe/js@2.8.21
+
+## 2.8.20
+
+### Patch Changes
+
+- @stackframe/js@2.8.20
+
 ## 2.8.19
 
 ### Patch Changes