+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/stack-shared
 
+## 2.8.21
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.20
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.19
 
 ### Patch Changes